Moved all GalleryCarousel functionality into the gallery overview component, moved carousel styling into a styled file
